Inline addBasicGame and createUUID
